requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
cupy-cuda11x>=12.0.0
pytest>=7.0.0
